+++ src/main/java/com/mss/hms/entities/Prescription.java
@@ -27,7 +27,7 @@ public class Prescription {
     private Date date;
 
     @NotEmpty(message = "prescription serial no can't be empty!!")
-    private Short serialNo;
+    private Integer serialNo;
 
     @NotEmpty(message = "prescription price can't be empty!!")
     private Double price;
